delay_time fix API 5 strcmp functions moved  to server find moved to server

+++ include/progs.h
@@ -230,7 +230,7 @@ typedef struct gedict_s {
 	char           *netname_non_team_drop;
 	int             all_active;
 	int             item_list;
-	int             delay_time;
+	float           delay_time;
 	int             dont_do_triggerwork;
 
 	int             numflames;
